Swap API projects
