+++ PFM.Api/PFM.API.sln
@@ -17,10 +17,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{5708C995-8
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PFM.Api.Contracts", "src\PFM.Api.Contracts\PFM.Api.Contracts.csproj", "{21255CEF-0816-45DE-B356-0544DE9D212A}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "v2", "v2", "{34DB9523-A612-4214-A696-63DAFADEF68F}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PFM.Api", "v2\PFM.Api\PFM.Api.csproj", "{B877F857-C5CD-4523-B1D3-2EC913733A85}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -47,10 +43,6 @@ Global
 		{21255CEF-0816-45DE-B356-0544DE9D212A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{21255CEF-0816-45DE-B356-0544DE9D212A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{21255CEF-0816-45DE-B356-0544DE9D212A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B877F857-C5CD-4523-B1D3-2EC913733A85}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B877F857-C5CD-4523-B1D3-2EC913733A85}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B877F857-C5CD-4523-B1D3-2EC913733A85}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B877F857-C5CD-4523-B1D3-2EC913733A85}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -61,7 +53,6 @@ Global
 		{3C0AB2B0-F26E-4D6F-8134-16A932544167} = {A2778004-ECF3-488C-AE9D-BC14AA293936}
 		{301FBC24-E53A-4378-9309-0DA5E2521679} = {5708C995-834F-4239-93B8-9AD8C6B24A68}
 		{21255CEF-0816-45DE-B356-0544DE9D212A} = {A2778004-ECF3-488C-AE9D-BC14AA293936}
-		{B877F857-C5CD-4523-B1D3-2EC913733A85} = {34DB9523-A612-4214-A696-63DAFADEF68F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {AA670247-C5F2-432F-9500-B1566565C984}
