+++ PFM.Api/src/PFM.Api/Controllers/AtmWithdrawController.cs
@@ -5,9 +5,9 @@ using PFM.Services.Interfaces;
 
 namespace PFM.Api.Controllers
 {
-    [Produces("application/json")]
-    [Route("api/AtmWithdraw")]
-    public class AtmWithdrawController : Controller
+    [ApiController]
+    [Route("api/[controller]")]
+    public class AtmWithdrawController : ControllerBase
     {
         public IAtmWithdrawService _atmWithdrawService { get; set; }
