+++ PFM.Api/src/PFM.Api/Controllers/BankAccountController.cs
@@ -5,9 +5,9 @@ using PFM.Services.Interfaces;
 
 namespace PFM.Api.Controllers
 {
-    [Produces("application/json")]
-    [Route("api/BankAccount")]
-    public class BankAccountController : Controller
+    [ApiController]
+    [Route("api/[controller]")]
+    public class BankAccountController : ControllerBase
     {
         private readonly IBankAccountService _BankAccountService;
