+++ PFM.Api/src/PFM.Api/Controllers/SalaryController.cs
@@ -5,9 +5,9 @@ using PFM.Services.Interfaces;
 
 namespace PFM.Api.Controllers
 {
-    [Produces("application/json")]
-    [Route("api/Salary")]
-    public class SalaryController : Controller
+    [ApiController]
+    [Route("api/[controller]")]
+    public class SalaryController : ControllerBase
     {
         private readonly ISalaryService _SalaryService;
