+++ PFM.Api/src/PFM.Api/Controllers/TaxTypeController.cs
@@ -5,9 +5,9 @@ using PFM.Services.Interfaces;
 
 namespace PFM.Api.Controllers
 {
-    [Produces("application/json")]
-    [Route("api/TaxType")]
-    public class TaxTypeController : Controller
+    [ApiController]
+    [Route("api/[controller]")]
+    public class TaxTypeController : ControllerBase
     {
         private readonly ITaxTypeService _TaxTypeService;
