+++ PFM.Api/src/PFM.Api/appsettings.Development.json
@@ -1,5 +1,3 @@
-﻿{
-  "ConnectionStrings": {
-    "PFMConnection": "Server=localhost,1433;Database=PFM_MAIN_DB;User Id=PFM_API_SVC;Password=Helloworld123!"
-  }
+{
+
 }
