+++ PFM.Api/src/PFM.Api/appsettings.json
@@ -1,4 +1,4 @@
-﻿{
+{
   "Serilog": {
     "Using": [
       "Serilog.Sinks.Console",
@@ -13,16 +13,14 @@
       }
     ],
     "Properties": {
-      "Application": "PFM.Api"
+      "Application": "PFM.Api.v2"
     }
   },
-  "ConnectionStrings": {
-    "PFMConnection": "Server=localhost,1433;Database=PFM_MAIN_DB;User Id=PFM_API_SVC;Password=Helloworld123!"
-  },
-  "JwtKey": "SOME_RANDOM_KEY_DO_NOT_SHARE",
-  "JwtIssuer": "http://yourdomain.com",
-  "JwtExpireDays": 30,
+  "AllowedHosts": "*",
   "AuthApi": {
     "EndpointUrl": "http://localhost:5000"
+  },
+  "ConnectionStrings": {
+    "PFMConnection": "Server=localhost,1433;Database=PFM_MAIN_DB;User Id=PFM_API_SVC;Password=Helloworld123!;TrustServerCertificate=True"
   }
 }
