--- PFM.Api/v2/PFM.Api/Middlewares/TimedOperationMiddleware.cs
@@ -1,39 +1,0 @@
-﻿using Microsoft.AspNetCore.Http;
-using System;
-using System.Net;
-using System.Threading.Tasks;
-
-namespace PFM.Api.Middlewares
-{
-    public class UnhandledExceptionMiddleware
-    {
-        private readonly RequestDelegate _next;
-        private readonly Serilog.ILogger _logger;
-
-        public UnhandledExceptionMiddleware(RequestDelegate next, Serilog.ILogger logger)
-        {
-            _logger = logger;
-            _next = next;
-        }
-
-        public async Task InvokeAsync(HttpContext httpContext)
-        {
-            try
-            {
-                await _next(httpContext);
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(ex, "Unhandled exception occurred");
-                await HandleExceptionAsync(httpContext);
-            }
-        }
-
-        private async Task HandleExceptionAsync(HttpContext context)
-        {
-            context.Response.ContentType = "application/json";
-            context.Response.StatusCode = (int)HttpStatusCode.InternalServerError;
-            await context.Response.WriteAsync($"Go check the application logs!");
-        }
-    }
-}
