--- PFM.Api/v2/PFM.Api/Middlewares/UnhandledExceptionMiddleware.cs
@@ -1,25 +1,0 @@
-﻿using Microsoft.AspNetCore.Http;
-using SerilogTimings;
-using System.Threading.Tasks;
-
-namespace PFM.Api.Middlewares
-{
-    public class TimedOperationMiddleware
-    {
-        private readonly RequestDelegate _next;
-
-        public TimedOperationMiddleware(RequestDelegate next)
-        {
-            _next = next;
-        }
-
-        public async Task InvokeAsync(HttpContext httpContext)
-        {
-            using (var op = Operation.Begin($"{httpContext.Request.Method} {httpContext.Request.Path}"))
-            {
-                await _next(httpContext);
-                op.Complete();
-            }
-        }
-    }
-}
