--- PFM.Api/v2/PFM.Api/Program.cs
@@ -1,60 +1,0 @@
-using Autofac;
-using Autofac.Extensions.DependencyInjection;
-using AutoMapper;
-using Microsoft.EntityFrameworkCore;
-using PFM.Api.Extensions;
-using PFM.Api.Middlewares;
-using PFM.DataAccessLayer;
-using PFM.Services.Core.Automapper;
-
-namespace PFM.Api
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            var builder = WebApplication.CreateBuilder(args);
-
-            builder.Services.AddControllers();
-
-            builder.Services.AddAuthenticationAndAuthorization(builder.Configuration);
-
-            builder.Services.AddMonitoring(builder.Configuration);
-
-            builder.Services.AddEndpointsApiExplorer();
-            builder.Services.AddSwaggerDefinition();
-
-            builder.Services.AddDbContext<PFMContext>(opts => opts.UseSqlServer(builder.Configuration.GetConnectionString("PFMConnection")));
-
-            builder.Host.UseServiceProviderFactory(new AutofacServiceProviderFactory());
-            builder.Host.ConfigureContainer<ContainerBuilder>(builder => builder.RegisterModule(new AutoFacModule()));
-
-            var app = builder.Build();
-
-            app.UseMiddleware<TimedOperationMiddleware>();
-            app.UseMiddleware<UnhandledExceptionMiddleware>();
-
-            if (app.Environment.IsDevelopment())
-            {
-                app.UseSwagger();
-                app.UseSwaggerUI();
-            }
-
-            app.UseHttpsRedirection();
-
-            app.UseAuthorization();
-
-            app.MapControllers();
-
-            Mapper.Initialize(cfg =>
-            {
-                cfg.AddProfile<ModelToEntityMapping>();
-                cfg.AddProfile<EntityToModelMapping>();
-                cfg.AddProfile<EntityToEntityMapping>();
-                cfg.AddProfile<SearchParametersMapping>();
-            });
-
-            app.Run();
-        }
-    }
-}
