--- PFM.Api/v2/PFM.Api/appsettings.Development.json
@@ -1,3 +1,0 @@
-{
-
-}
