--- PFM.Api/v2/PFM.Api/appsettings.json
@@ -1,26 +1,0 @@
-{
-  "Serilog": {
-    "Using": [
-      "Serilog.Sinks.Console",
-      "Serilog.Sinks.Seq"
-    ],
-    "MinimumLevel": "Debug",
-    "WriteTo": [
-      { "Name": "Console" },
-      {
-        "Name": "Seq",
-        "Args": { "serverUrl": "http://localhost:5341" }
-      }
-    ],
-    "Properties": {
-      "Application": "PFM.Api.v2"
-    }
-  },
-  "AllowedHosts": "*",
-  "AuthApi": {
-    "EndpointUrl": "http://localhost:5000"
-  },
-  "ConnectionStrings": {
-    "PFMConnection": "Server=localhost,1433;Database=PFM_MAIN_DB;User Id=PFM_API_SVC;Password=Helloworld123!;TrustServerCertificate=True"
-  }
-}
